Rename class OrganizationFolderTest, craete testRenameOrganizationFolder(), modified testCreateOrganizationFolder() locators
